+++ ipapp/db/oracle.py
@@ -5,6 +5,7 @@ https://cx-oracle.readthedocs.io/en/latest/user_guide/introduction.html
 
 https://container-registry.oracle.com/pls/apex/f?p=113:4:103772846430850::NO:::
 """
+
 import asyncio
 import functools
 from contextvars import Token