+++ ipapp/openapi/models.py
@@ -247,9 +247,9 @@ class PathItem(BaseModel):
 
 # Workaround OpenAPI recursive reference
 class OperationWithCallbacks(BaseModel):
-    callbacks: Optional[
-        Dict[str, Union[Dict[str, PathItem], Reference]]
-    ] = None
+    callbacks: Optional[Dict[str, Union[Dict[str, PathItem], Reference]]] = (
+        None
+    )
 
 
 class SecuritySchemeType(Enum):
@@ -335,13 +335,13 @@ class Components(BaseModel):
     examples: Optional[Dict[str, Union[Example, Reference]]] = None
     requestBodies: Optional[Dict[str, Union[RequestBody, Reference]]] = None
     headers: Optional[Dict[str, Union[Header, Reference]]] = None
-    securitySchemes: Optional[
-        Dict[str, Union[SecurityScheme, Reference]]
-    ] = None
+    securitySchemes: Optional[Dict[str, Union[SecurityScheme, Reference]]] = (
+        None
+    )
     links: Optional[Dict[str, Union[Link, Reference]]] = None
-    callbacks: Optional[
-        Dict[str, Union[Dict[str, PathItem], Reference]]
-    ] = None
+    callbacks: Optional[Dict[str, Union[Dict[str, PathItem], Reference]]] = (
+        None
+    )
 
 
 class Tag(BaseModel):