+++ tests/test_mq_pika.py
@@ -18,7 +18,7 @@ from ipapp.mq.pika import (
 async def wait_for(fn: Callable, timeout=60) -> Any:
     err = None
     try:
-        with async_timeout(timeout):
+        async with async_timeout(timeout):
             while True:
                 try:
                     if asyncio.iscoroutine(fn):
@@ -39,7 +39,7 @@ async def wait_for(fn: Callable, timeout=60) -> Any:
             raise
 
 
-async def test_pika(loop, rabbitmq_url):
+async def test_pika(rabbitmq_url):
 
     messages: List[Tuple[bytes]] = []
 
@@ -59,7 +59,7 @@ async def test_pika(loop, rabbitmq_url):
             await self.consume('myqueue', self.message)
 
         async def message(
-            self, body: bytes, deliver: Deliver, proprties: Properties
+            self, body: bytes, deliver: Deliver, properties: Properties
         ) -> None:
             await self.ack(delivery_tag=deliver.delivery_tag)
             messages.append((body,))
@@ -86,9 +86,8 @@ async def test_pika(loop, rabbitmq_url):
     await app.stop()
 
 
-async def test_dead_letter_exchange(loop, rabbitmq_url):
+async def test_dead_letter_exchange(rabbitmq_url):
     messages: List[Tuple[bytes]] = []
-    queue: str
 
     class TestPubChg(PikaChannel):
         name = 'pub'
@@ -123,10 +122,10 @@ async def test_dead_letter_exchange(loop, rabbitmq_url):
             await self.consume('myqueue1', self.message)
 
         async def message(
-            self, body: bytes, deliver: Deliver, proprties: Properties
+            self, body: bytes, deliver: Deliver, properties: Properties
         ) -> None:
             await self.ack(delivery_tag=deliver.delivery_tag)
-            messages.append((body, proprties))
+            messages.append((body, properties))
 
     app = BaseApplication(BaseConfig())
     app.add(